Squash common things into kv::Encryptor
